Customize the tool title
Customized the tool title in both the screen output and in the generated SPDX to distinguish the output of this tool from the original.

Also, added the version of the tool to the SPDX file in case there are bugs we'll have an easy way to determine what version of the tool was used to generate the incorrect file.

Resolves #7